Add nextjs catch all error page
