+++ src/app/error/page.tsx
@@ -1,27 +1,7 @@
 'use client';
 
-import { ArrowLeftIcon } from 'lucide-react';
-import Link from 'next/link';
-import { useSearchParams } from 'next/navigation';
+import NextErrorPage from '@/app/error';
 
 export default function ErrorPage() {
-  const searchParams = useSearchParams();
-  const error = searchParams.get('error') || 'An unknown error occurred';
-
-  return (
-    <div className="hero bg-base-200 container mx-auto mt-4">
-      <div className="hero-content min-h-96 text-center">
-        <div className="max-w-md">
-          <h1 className="text-5xl font-bold text-red-500">Oops!</h1>
-          <p className="py-6 text-red-500">{error}</p>
-
-          <Link href="/" passHref>
-            <button className="btn btn-accent">
-              Home <ArrowLeftIcon />
-            </button>
-          </Link>
-        </div>
-      </div>
-    </div>
-  );
+  return <NextErrorPage />;
 }
